refactor for easier testing, added state classes
