+++ EscapeVR/src/main/java/nl/tudelft/kroket/state/State.java
@@ -1,0 +1,18 @@
+package nl.tudelft.kroket.state;
+
+import nl.tudelft.kroket.escape.SetUp;
+import nl.tudelft.kroket.escape.Launcher;
+
+public abstract class State {
+  
+  public State(){
+    
+  }
+  
+  public abstract void update(float tpf);
+  
+  public abstract void stopState();
+  
+  public abstract void startState();
+  
+}
