+++ EscapeVR/src/main/java/nl/tudelft/kroket/state/states/NoState.java
@@ -1,0 +1,35 @@
+package nl.tudelft.kroket.state.states;
+
+import nl.tudelft.kroket.escape.SetUp;
+import nl.tudelft.kroket.escape.Launcher;
+import nl.tudelft.kroket.state.State;
+
+public class NoState extends State{
+
+  public NoState(){
+
+  }
+
+  public void update(float tpf){
+    if (Launcher.mainApplication.getSetUp() != null) {
+
+    }   
+  }
+
+  @Override
+  public void stopState() {
+    if (Launcher.mainApplication.getSetUp() != null) {
+      
+    }
+    
+  }
+
+  @Override
+  public void startState() {
+    if (Launcher.mainApplication.getSetUp() != null) {
+      
+    }
+    
+  }
+
+}
