+++ EscapeVR/src/test/java/nl/tudelft/kroket/screen/LobbyScreenTest.java
@@ -1,0 +1,14 @@
+package nl.tudelft.kroket.screen;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class LobbyScreenTest {
+
+  @Test
+  public void test() {
+    fail("Not yet implemented");
+  }
+
+}
