+++ EscapeVR/src/test/java/nl/tudelft/kroket/screen/ScreenManagerTest.java
@@ -1,0 +1,19 @@
+package nl.tudelft.kroket.screen;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+/**
+ * Class for testing the ScreenManager class.
+ * @author Kroket
+ *
+ */
+public class ScreenManagerTest {
+
+  @Test
+  public void test() {
+    fail("Not yet implemented");
+  }
+
+}
